move fixtures.ts outside fixtures dir in order for babel to compile it
