+++ components/legacy/e2e-helper/e2e-fs-helper.ts
@@ -3,7 +3,7 @@ import chai, { expect } from 'chai';
 import glob from 'glob';
 import * as path from 'path';
 import { generateRandomStr } from '@teambit/toolbox.string.random';
-import * as fixtures from './fixtures/fixtures';
+import * as fixtures from './fixtures';
 import { ensureAndWriteJson } from './e2e-helper';
 import ScopesData from './e2e-scopes';
